Add TestOtimization: BinaryExpression with LiteralInt and add unitary Test: toString and isBooleanTrue
