+++ liquidjava-verifier/src/test/java/liquidjava/rj_language/TestLiteralString.java
@@ -1,6 +1,14 @@
 package liquidjava.rj_language;
 
+import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotEquals;
+import static org.junit.Assert.assertNotSame;
+import static org.junit.Assert.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+import java.io.PrintStream;
+
 import org.junit.Test;
 import liquidjava.rj_language.ast.LiteralString;
 
@@ -14,4 +22,24 @@ public class TestLiteralString {
         // Act & Assert: Executar e verificar
         assertNotEquals(s1.hashCode(), s2.hashCode());
     }
+
+    @Test
+    public void testToString() {
+        LiteralString myString = new LiteralString("olá");
+
+        String resultado = myString.toString();
+
+        assertEquals("olá", resultado);
+    }
+
+    @Test
+    public void isBooleanTrue() {
+        LiteralString myString = new LiteralString("yes");
+
+        boolean result = myString.isBooleanTrue();
+
+        assertFalse(result);
+
+    }
+
 }
